Enable drag lock on touchpads
Somehow the default setting was changed to enabled_sticky in openSUSE. I
find that terribly inconvenient. Setting it to enabled (I think it was
disabled before) is actually quite convenient as it gives a short but
useful time window to lift your finger to get more space on the trackpad
while dragging.